Log Store further implemented
The log store activity now shows a scrollable view of all current logs,
which can be deleted at the press of a button. Currently, it won't
refresh until you exit and go back to the page. I'll fix that soon.
Additionally, I changed the icon to a picture of Evan's head.