Added Log Location Warning
I added a warning box that comes up if the location is not enabled on
the phone. It needs to be moved to another location in the code, so it
doesn't come up a million times in a row. I also made it so the erase
logs button works instantly. Also, we're still not sure if the speed is
actually being recorded or not. According to some sources, it may take a
few minutes for a valid GPS signal to be found.