Increased Functionality to Log Storage
Now, you can select the file from the list of logs to view the contents
of that log. I consider this to be a very large step because now we have
the *most basic core functionality* of the app working, and we can
extend from here. Of course, this may be revised slightly, but it
currently does work.