Changes to the Setting GUI and More.
I placed all the widgets in the setting activity in a scroll view. Now
all the information in the settings activity can now be seen on any
phone no matter what the size of the screen. I also began implementing
the User Defined Speed system. I have basic implementation of a popup
window done but there is still more work to be done.